consolidate scores attributes component
